requirements.txt
pandas==1.5.3
numpy==1.24.3
sqlalchemy==1.4.53
psycopg2-binary==2.9.7
python-dotenv==1.0.0
paramiko==3.2.0
sshtunnel==0.4.0